requirements.txt
tensorflow
numpy
opencv-python
scikit-learn
pyserial
